[FEATURE] Configure Tempo data source in UI
* extract HTTP Settings editor from PrometheusDatasourceEditor
* for Prometheus, keep the "scrape interval" setting when switching
  between direct and proxy mode
* add TempoDatasourceEditor
