+++ ui/core/src/model/http.ts
@@ -11,4 +11,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { HTTPProxy } from './http-proxy';
+
 export type RequestHeaders = Record<string, string>;
+
+export interface HTTPDatasourceSpec {
+  directUrl?: string;
+  proxy?: HTTPProxy;
+}
